Refactor SettingsOverlay component to be a functional one.
